feat: extracted out admin nav
