+++ app/(admin)/layout.tsx
@@ -2,27 +2,17 @@
 
 import React, { useEffect, useState } from "react";
 import styles from "../../styles/AdminDashboard.module.css";
-import { onAuthStateChanged, signOut, User } from "firebase/auth";
+import { onAuthStateChanged, User } from "firebase/auth";
 import { auth } from "../lib/firebase";
-import { useRouter, usePathname } from "next/navigation";
-import Link from "next/link";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-  faTachometerAlt,
-  faCalendarAlt,
-  faClock,
-  faUsers,
-  faConciergeBell,
-  faSignOutAlt,
-} from "@fortawesome/free-solid-svg-icons";
+import { useRouter } from "next/navigation";
+import DashboardNav from "../components/DashboardNav";
 
-export default function AdminLayout({
-  children,
-}: Readonly<{
+interface AdminLayoutProps {
   children: React.ReactNode;
-}>) {
+}
+
+const AdminLayout: React.FC<AdminLayoutProps> = ({ children }) => {
   const navigation = useRouter();
-  const pathname = usePathname();
   const [user, setUser] = useState<User | null>(null);
   const [loading, setLoading] = useState(true);
 
@@ -39,15 +29,6 @@ export default function AdminLayout({
     return () => unsubscribe();
   }, [navigation]);
 
-  const handleSignOut = async () => {
-    try {
-      await signOut(auth);
-      navigation.push("/login");
-    } catch (error) {
-      console.error("Sign out error:", error);
-    }
-  };
-
   if (loading) {
     return <p className={styles.loadingText}>Loading...</p>;
   }
@@ -55,57 +36,11 @@ export default function AdminLayout({
   return user ? (
     <>
       <main className={styles.dashboardMain}>
-        <nav className={styles.dashboardNav}>
-          <h1 className={styles.dashboardTitle}>Admin Dashboard</h1>
-          <ul>
-            <Link href={"/admin"}>
-              <li className={pathname === "/admin" ? styles.activeLink : ""}>
-                <FontAwesomeIcon icon={faTachometerAlt} /> Dashboard
-              </li>
-            </Link>
-            <Link href={"/admin/appointments"}>
-              <li
-                className={
-                  pathname === "/admin/appointments" ? styles.activeLink : ""
-                }
-              >
-                <FontAwesomeIcon icon={faCalendarAlt} /> Manage Appointments
-              </li>
-            </Link>
-            <Link href={"/admin/clients"}>
-              <li
-                className={
-                  pathname === "/admin/clients" ? styles.activeLink : ""
-                }
-              >
-                <FontAwesomeIcon icon={faUsers} /> Clients
-              </li>
-            </Link>
-            <Link href={"/admin/services"}>
-              <li
-                className={
-                  pathname === "/admin/services" ? styles.activeLink : ""
-                }
-              >
-                <FontAwesomeIcon icon={faConciergeBell} /> Services
-              </li>
-            </Link>
-            <Link href={"/admin/workinghours"}>
-              <li
-                className={
-                  pathname === "/admin/workinghours" ? styles.activeLink : ""
-                }
-              >
-                <FontAwesomeIcon icon={faClock} /> Working Hours
-              </li>
-            </Link>
-          </ul>
-          <button onClick={handleSignOut} className={styles.button}>
-            <FontAwesomeIcon icon={faSignOutAlt} /> Sign Out
-          </button>
-        </nav>
+        <DashboardNav />
         <section className={styles.dashboardContent}>{children}</section>
       </main>
     </>
   ) : null;
-}
+};
+
+export default AdminLayout;
